did some cleanup for markket.api/store
